LoadSavedModel
Changes to save the entire model and load it instead of only state dict.

This has the advantage of not needing to track the folder to net structure mapping for reconstructing net, however if code is refactored it could cause trouble. (Technically if code is refactored even loading state dict might not work depending on the changes)