Policy Evaluation
First stab at policy evaluation.

Code is a little fragile because rollout logic depends on how dynamics and policy are trained ... (the inputs need to be processed according to that)

For now, the eval code is for a specific policy and dynamics networks (the most likely case), as we see a need for more generic evaluation, we should evolve code accordingly.